program/site/ml_api_project/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pillow==10.4.0

# PyTorch / Torchvision (для macOS ставляться pip'ом без CUDA)
torch==2.4.1
torchvision==0.19.1
